add entity and remove client.

+++ app/src/main/java/me/aluceps/repofinder/data/api/response/Repositories.kt
@@ -3,9 +3,9 @@ package me.aluceps.repofinder.data.api.response
 import com.google.gson.annotations.SerializedName
 
 data class Repositories(
-    @SerializedName("total_count")
-    val count: Int,
-    @SerializedName("incomplete_results")
-    val incomplete: Boolean,
-    val items: List<Repository>
+        @SerializedName("total_count")
+        val count: Long,
+        @SerializedName("incomplete_results")
+        val incomplete: Boolean,
+        val items: List<Repository>
 )
